# src/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "REST API proyectoUTM",
        "description": "<html><body>Esta es la documentación de la API para el sistema informatico.<p>Para poder usar esta API es necesario solicitar el token y con esto poder utilizar los servicios proporcionados.</p><br><p><b>Autenticación y token de API</b></p><p>El sistema proporciona elapi-token. Una vez que haya creado y confirmado su cuenta.</p></body></html>",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "host": "localhost:3000",
    "basePath": "",
...
    "schemes": [
        "http"
    ],
    "consumes":[
        "application/json"
    ],
    "produces" :[
        "application/json"
    ],
    "paths":
...
        "/api/carritos/cancelarcarrito":{
            "post":{
                "summary": "Servicio para cancelar un carrito de compras",
                "operationId":"cancelar",
                "produces":[
                    "application/json"
                ],
                "parameters":[
                    {
                        "name":"Cancelar un carrito",
                        "in": "body",
                        "description":"Este servicio nos permite poner en un carrito existente el estado de cancelado, para que no se pueda realizar la compra ademas de que todos los productos que se habian agregado al carrito se devuelven a la cantidad que tenian antes de que se agregaran al carrito",
                        "requiered":true,
                        "schema":{
                            "$ref":"#/definitions/CancelarCarrito"
                        } 
                    }
                ],
                "tags":["Carritos"],
                "responses":{
                    "200":{
                        "description":"respuesta correcta",
                        "examples": {
                            "application/json": [
                                {
                                    "idCarrito":1
                                } 
                            ]
                        }
                    },
                    "403":{
                        "description":"Prohibido"
                    }
                }
            }
        },
...
    },
    "definitions":{
...
        "CancelarCarrito":{
            "type": "object",
            "properties": {
                "id": {
                    "type": "number",
                    "description": "Codigo que permite identificar el carrito"
                }
            }
        },
...
    }
}